Storage: remove API inconsistency
Retrieval of non-existent bird would return None, but attempt to remove
it would raise an exception. In the world of web APIs with many
concurrent clients having already-removed ID will sometimes happen, so
that is really not exceptional enough to raise an exception.

StorageEngine#remove will now return True if object was removed, and
False, if it was not there.